requirements.txt
streamlit==1.42.0
pandas==2.2.3
numpy==2.1.3
openpyxl==3.1.5
gspread==6.1.2
google-auth==2.37.0
google-api-python-client==2.162.0
